33-mamadou-field-on-map

+++ backend/static_map.py
@@ -1,3 +1,4 @@
+import os
 import requests
 
 # URLs antiguas
@@ -7,17 +8,16 @@ tmp_url = "https://maps.googleapis.com/maps/api/staticmap?center={},{}&zoom={}&s
 def crea_url(data):
     return tmp_url.format(data["lat"], data["lon"], data["zoom"], data["key"])  
 
-def save_imagen(data, image_name):
+def save_imagen(data, image_name, overwrite=False):
     url = crea_url(data)
-    r = requests.get(url)
-    f = open('%s.png' % image_name,'wb')
-    f.write(r.content)
-    f.close()
-
-# proxy = ""  # In case you need a proxy
-# proxies = {"http": proxy, "https": proxy}
-# sesion = requests.Session()
-# sesion.proxies = proxies
+    if overwrite==True and not os.path.exists(image_name) :
+        r = requests.get(url)
+        f = open('%s.png' % image_name,'wb')
+        f.write(r.content)
+        f.close()
+        print(f"{image_name} created !")
+    else :
+        print(f"{image_name} exists !")
 
 # Aqui se modifican los datos
 data = {"lat": 6.755331569175543, "lon": -4.530004939961217,
